:recycle: Update inheritance to remove duplicate code
