+++ src/openpersonen/api/views/base/__init__.py
@@ -1,0 +1,3 @@
+from .base_viewset import *  # noqa
+from .historie_viewset import *  # noqa
+from .nested_viewset import *  # noqa
