+++ src/openpersonen/api/views/kind.py
@@ -1,42 +1,10 @@
-from rest_framework.permissions import IsAuthenticated
-from rest_framework.response import Response
-from rest_framework.status import HTTP_200_OK
-from rest_framework.viewsets import ViewSet
-
 from openpersonen.api.data_classes import Kind
 from openpersonen.api.serializers import KindSerializer
+from openpersonen.api.views.base import NestedViewSet
 
 
-class KindViewSet(ViewSet):
+class KindViewSet(NestedViewSet):
 
     lookup_field = "id"
     serializer_class = KindSerializer
-    permission_classes = [IsAuthenticated]
-
-    def get_serializer_context(self):
-        """
-        Extra context provided to the serializer class.
-        """
-        return {"request": self.request, "format": self.format_kwarg, "view": self}
-
-    def get_serializer(self, *args, **kwargs):
-        """
-        Return the serializer instance that should be used for validating and
-        deserializing input, and for serializing output.
-        """
-        serializer_class = self.serializer_class
-        kwargs["context"] = self.get_serializer_context()
-        return serializer_class(*args, **kwargs)
-
-    def list(self, request, *args, **kwargs):
-        return Response(data=[], status=HTTP_200_OK)
-
-    def retrieve(self, request, *args, **kwargs):
-        burgerservicenummer = kwargs["ingeschrevenpersonen_burgerservicenummer"]
-        id = kwargs["id"]
-
-        instance = Kind.retrieve(burgerservicenummer)
-
-        serializer = self.serializer_class(instance)
-
-        return Response(data=serializer.data, status=HTTP_200_OK)
+    instance_class = Kind
